% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucnsim_functions.R
\name{estimate_transition_rates}
\alias{estimate_transition_rates}
\title{Estimate status transition rates}
\usage{
estimate_transition_rates(
  extant_taxa_current_status,
  iucn_history_file,
  outdir,
  possibly_extinct_list,
  extinction_probs_mode = 0,
  species_specific_regression = FALSE,
  rate_samples = 100,
  n_gen = 1e+05,
  burnin = 1000,
  seed = NULL,
  load_from_file = FALSE,
  return_object = TRUE
)
}
\description{
Estimate status transition rates
}
