% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucnn_functions.R
\name{get_invalid_statuses}
\alias{get_invalid_statuses}
\title{IUCNN functions to estimate missing statuses using neural network}
\usage{
get_invalid_statuses(current_status_df)
}
\description{
IUCNN functions to estimate missing statuses using neural network
}
