% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_future_sim.R
\name{run_future_sim}
\alias{run_future_sim}
\title{Simulate future extinctions}
\usage{
run_future_sim(
  simulation_input_data,
  outdir,
  n_years = 100,
  n_sim = 10000,
  status_change = TRUE,
  conservation_increase_factor = 1,
  threat_increase_factor = 1,
  model_unknown_as_lc = FALSE,
  until_n_taxa_extinct = 0,
  plot_diversity_trajectory = TRUE,
  plot_status_trajectories = TRUE,
  plot_histograms = FALSE,
  plot_status_piechart = TRUE,
  seed = NULL,
  load_from_file = FALSE
)
}
\description{
Simulate future extinctions
}
